+++ scout/utils/ensembl_rest_clients.py
@@ -1,20 +1,22 @@
 # -*- coding: UTF-8 -*-
 import json
 import logging
+import requests
 from urllib.error import HTTPError
 from urllib.parse import urlencode
 from urllib.request import Request, urlopen
 
 LOG = logging.getLogger(__name__)
 
-SERVER_37 = 'https://grch37.rest.ensembl.org'
-SERVER_38 = 'http://rest.ensembl.org/'
 HEADERS = {'Content-type':'application/json'}
-BIOMART =  "http://www.ensembl.org/biomart/martservice?query="
-
+RESTAPI_37 = 'http://grch37.rest.ensembl.org'
+RESTAPI_38 = 'http://rest.ensembl.org/'
 PING_ENDPOINT = 'info/ping'
 
-class EnsemblRestClient:
+BIOMART_37 =  "http://grch37.ensembl.org/biomart/martservice?query="
+BIOMART_38 =  "http://ensembl.org/biomart/martservice?query="
+
+class EnsemblRestApiClient:
     """A class handling requests and responses to and from the Ensembl REST APIs.
     Endpoints for human build 37: https://grch37.rest.ensembl.org
     Endpoints for human build 38: http://rest.ensembl.org/
@@ -24,11 +26,11 @@ class EnsemblRestClient:
 
     def __init__(self, build='37'):
         if build == '38':
-            self.server = SERVER_38
+            self.server = RESTAPI_38
         else:
-            self.server = SERVER_37
+            self.server = RESTAPI_37
 
-    def ping_server(self, server=SERVER_37):
+    def ping_server(self, server=RESTAPI_38):
         """ping ensembl
 
         Accepts:
@@ -58,7 +60,7 @@ class EnsemblRestClient:
             endpoint += '?' + urlencode(params)
 
         url = ''.join([ self.server, endpoint])
-        LOG.info('Using Ensembl API with the url:{}'.format(url))
+        LOG.info('Using Ensembl API with the following url:{}'.format(url))
         data = self.send_request(url)
         return data
 
@@ -86,56 +88,89 @@ class EnsemblRestClient:
             data = e
         return data
 
-    def create_biomart_xml(self, filters, attributes):
+
+class EnsemblBiomartClient:
+
+    def __init__(self, build='37'):
+        if build == '38':
+            self.server = BIOMART_38
+        else:
+            self.server = BIOMART_37
+
+    def query_service(self, xml, temp_file):
+        """Query the Ensembl biomart service
+
+        Accepts:
+            xml(str): an xml formatted query, as described here:
+                https://grch37.ensembl.org/info/data/biomart/biomart_perl_api.html
+
+            temp_file(NamedTemporaryFile)
+        """
+        url = ''.join([self.server, xml])
+        try:
+            with requests.get(url, stream=True) as r:
+                with open(temp_file.name, 'wb') as f:
+                    for chunk in r.iter_content(chunk_size=8192):
+                        if chunk:
+                            f.write(chunk)
+        except Exception as ex:
+            LOG.info('Error downloading data from biomart: {}'.format(ex))
+            return ex
+
+        return
+
+
+    def create_biomart_xml(self, filters=None, attributes=None):
         """Convert biomart query params into biomart xml query
 
         Accepts:
-            filters(list): a list of tuples with filter name and values (ex: ('chromosome_name' : [1,2,'X']) )
+            filters(list): a list of tuples with filter name and values, ex: ('chromosome_name' , [1,2,'X'])
             attributes(list): a list of attributes
 
         Returns:
             xml: a query xml file
 
         """
-        filter_lines = [ self.xml_filter(filter[0], filter[1]) for filter in filters]
-        attribute_lines = [ self.xml_attribute(attr_name) for attr_name in attributes ]
+        filter_lines = self.xml_filters(filters)
+        attribute_lines = self.xml_attributes(attributes)
         xml = """<?xml version="1.0" encoding="UTF-8"?>
         <!DOCTYPE Query>
-        <Query  virtualSchemaName = "default" formatter = "TSV" header = "0" uniqueRows = "0" count = "" datasetConfigVersion = "0.6" >
-
+        <Query  virtualSchemaName = "default" formatter = "TSV" header = "0" uniqueRows = "0" count = "" datasetConfigVersion = "0.6" completionStamp = "1">
         	<Dataset name = "hsapiens_gene_ensembl" interface = "default" >
                 {0}
                 {1}
         	</Dataset>
-        </Query>""".format( filter_lines, attribute_lines)
+        </Query>""".format( filter_lines,attribute_lines)
 
         return xml
 
-    def xml_filter(self, name, value):
+    def xml_filters(self, filters):
         """Creates a filter line for the biomart xml document
 
         Accepts:
-            name(str): filter name
-            value(str or list): filter value
+            filters(dict): keys are filter names and values are filter values
 
         Returns:
-            formatted_line(str): formatted xml line
+            formatted_lines(str): formatted xml line
         """
-        formatted_line = ""
-        if type(value) == list:
-            formatted_line='<Filter name = "{0}" value = "{1}"/>'.format(name, ','.join(str(value)))
-        else:
-            formatted_line='<Filter name = "{0}" value = "{1}"/>'.format(name, value)
-        return formatted_line
+        formatted_lines = ""
+        for key, value in filters.items():
+            if type(value) == list:
+                formatted_lines += '<Filter name = "{0}" value = "{1}"/>'.format(key, ','.join(value))
+            else:
+                formatted_lines += '<Filter name = "{0}" value = "{1}"/>'.format(key, value)
+        return formatted_lines
 
-    def xml_attribute(self, name):
+    def xml_attributes(self, attributes):
         """Creates an attribute line for the biomart xml document
 
         Accepts:
-            name(str): attribute name
+            attributes(list): attribute name
 
         Returns:
-            formatted_line(str): formatted xml line
+            formatted_lines(str): formatted xml line
         """
-        formatted_line = '<Attribute name = "{}" />'.format(name)
-        return formatted_line
+        formatted_lines = ""
+        for attr in attributes:
+            formatted_lines += '<Attribute name = "{}" />'.format(attr)
+        return formatted_lines